ml-backend/services/clothAnalyzer/requirements.txt
pytest
uvicorn
python-dotenv
fastapi
opencv-python
requests
openai